DI_Study/DASH_JIRA/DASH_JIRA_2/requirements.txt
psycopg2-binary
pydantic
pydantic-settings
openai
pandas
